highly optmized SNES_multiplayer class
[corrected]

now we're really talking SNES console language, as we read multiple controller using the same line and signals for clock and latch, and reading each bit of all controllers at the same time (and only then clocking controller line)

SNES controller protocol is not costly for arduino. but with naive implementation of multiplayer (read one after other in a for) on the SNES it caused controllers to be latched N times for each reading cycle (N = number of controllers), since all controllers share the same latch and clock line

(with 4 controllers (and my initial midi interface draft), we notice some delay. now i don't have the 4 controllers anymore, i have only 2)

first gametest with 2 controllers on ZSNES was perfect. the reading rate of this example must be way faster than original console.